Fix EvaluateCanExecute for find all implementations, closes #3045
